Refined features

+++ domain/src/main/java/com/himanshu_kumar/domain/usecase/CartSummaryUseCase.kt
@@ -3,5 +3,5 @@ package com.himanshu_kumar.domain.usecase
 import com.himanshu_kumar.domain.repository.CartRepository
 
 class CartSummaryUseCase(private val cartRepository:CartRepository) {
-    suspend fun execute(userId:Int) = cartRepository.getSummary(userId)
+    suspend fun execute(userId:Long) = cartRepository.getSummary(userId)
 }
